pyproject.toml
[project]
name = "django-htmx-starter"
version = "0.1.0"
description = "Django + HTMX + TailwindCSS starter template for hypermedia applications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-htmx>=1.21.0",
    "django-allauth[socialaccount]>=65.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=5.0.0",
]
prod = [
    "gunicorn>=23.0.0",
]

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=5.0.0",
]
